model-services/timesfm/requirements.txt
fastapi
uvicorn
pydantic
torch
python-dateutil
timesfm[torch]
numpy<2.0